+++ plugins/inputs/prometheus/parser.go
@@ -1,28 +1,61 @@
-package v1
+package prometheus
 
 import (
+	"bufio"
+	"bytes"
 	"fmt"
+	"io"
 	"math"
+	"mime"
+	"net/http"
 	"time"
 
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/metric"
 	. "github.com/influxdata/telegraf/plugins/parsers/prometheus/common"
 
+	"github.com/matttproud/golang_protobuf_extensions/pbutil"
 	dto "github.com/prometheus/client_model/go"
+	"github.com/prometheus/common/expfmt"
 )
 
-// Parse returns a slice of Metrics from a text representation of a
-// metrics
-func Parse(metricFamilies map[string]*dto.MetricFamily, defaultTags map[string]string, now time.Time) ([]telegraf.Metric, error) {
+func Parse(buf []byte, header http.Header) ([]telegraf.Metric, error) {
+	var parser expfmt.TextParser
 	var metrics []telegraf.Metric
 	var err error
+	// parse even if the buffer begins with a newline
+	buf = bytes.TrimPrefix(buf, []byte("\n"))
+	// Read raw data
+	buffer := bytes.NewBuffer(buf)
+	reader := bufio.NewReader(buffer)
 
+	// Prepare output
+	metricFamilies := make(map[string]*dto.MetricFamily)
+
+	if isProtobuf(header) {
+		for {
+			mf := &dto.MetricFamily{}
+			if _, ierr := pbutil.ReadDelimited(reader, mf); ierr != nil {
+				if ierr == io.EOF {
+					break
+				}
+				return nil, fmt.Errorf("reading metric family protocol buffer failed: %s", ierr)
+			}
+			metricFamilies[mf.GetName()] = mf
+		}
+	} else {
+		metricFamilies, err = parser.TextToMetricFamilies(reader)
+		if err != nil {
+			return nil, fmt.Errorf("reading text format failed: %s", err)
+		}
+	}
+
+	now := time.Now()
 	// read metrics
 	for metricName, mf := range metricFamilies {
 		for _, m := range mf.Metric {
 			// reading tags
-			tags := MakeLabels(m, defaultTags)
+			tags := MakeLabels(m, nil)
 
 			// reading fields
 			fields := make(map[string]interface{})
@@ -58,6 +91,18 @@ func Parse(metricFamilies map[string]*dto.MetricFamily, defaultTags map[string]s
 	}
 
 	return metrics, err
+}
+
+func isProtobuf(header http.Header) bool {
+	mediatype, params, error := mime.ParseMediaType(header.Get("Content-Type"))
+
+	if error != nil {
+		return false
+	}
+
+	return mediatype == "application/vnd.google.protobuf" &&
+		params["encoding"] == "delimited" &&
+		params["proto"] == "io.prometheus.client.MetricFamily"
 }
 
 // Get Quantiles from summary metric
