+++ plugins/inputs/prometheus/prometheus.go
@@ -15,7 +15,7 @@ import (
 	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/plugins/common/tls"
 	"github.com/influxdata/telegraf/plugins/inputs"
-	parser "github.com/influxdata/telegraf/plugins/parsers/prometheus"
+	parser_v2 "github.com/influxdata/telegraf/plugins/parsers/prometheus"
 )
 
 const acceptHeader = `application/vnd.google.protobuf;proto=io.prometheus.client.MetricFamily;encoding=delimited;q=0.7,text/plain;version=0.0.4;q=0.3,*/*;q=0.1`
@@ -322,14 +322,11 @@ func (p *Prometheus) gatherURL(u URLAndAddress, acc telegraf.Accumulator) error 
 		return fmt.Errorf("error reading body: %s", err)
 	}
 
-	isprotobuf, err := parser.IsProtobuf(resp.Header)
-
-	if err == nil {
-		parser := parser.Parser{
-			MetricVersion: p.MetricVersion,
-			Protobuf:      isprotobuf,
-		}
+	if p.MetricVersion == 2 {
+		parser := parser_v2.Parser{}
 		metrics, err = parser.Parse(body)
+	} else {
+		metrics, err = Parse(body, resp.Header)
 	}
 
 	if err != nil {
