+++ plugins/parsers/registry.go
@@ -402,8 +402,6 @@ func NewFormUrlencodedParser(
 
 func NewPrometheusParser(defaultTags map[string]string) (Parser, error) {
 	return &prometheus.Parser{
-		DefaultTags:   defaultTags,
-		MetricVersion: 2,
-		Protobuf:      false,
+		DefaultTags: defaultTags,
 	}, nil
 }
